requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-ninja==1.2.1
pydantic==2.8.2
pydantic_core==2.20.1
sqlparse==0.5.0
typing_extensions==4.12.2
